+++ crates/bcr-wallet-ffi/src/api/mod.rs
@@ -1,4 +1,6 @@
-use bcr_wallet_core::types::{get_btc_tx_id, get_payment_type, get_transaction_status};
+use bcr_wallet_core::types::{
+    get_btc_alpha_tx_id, get_btc_beta_tx_id, get_payment_type, get_transaction_status,
+};
 use nostr_sdk::RelayUrl;
 use once_cell::sync::Lazy;
 use std::{panic, path::PathBuf, str::FromStr, sync::Arc, time::Duration};
@@ -788,15 +790,22 @@ pub struct Transaction {
     pub memo: Option<String>,
     pub ptype: PaymentType,
     pub status: TransactionStatus,
-    pub btc_tx_id: Option<String>,
+    pub melt_tx: MeltTx,
     pub quote_id: Option<String>,
+}
+
+#[derive(Debug, Clone)]
+pub struct MeltTx {
+    pub alpha_tx_id: Option<String>,
+    pub beta_tx_id: Option<String>,
 }
 
 impl std::convert::From<cdk::wallet::types::Transaction> for Transaction {
     fn from(tx: cdk::wallet::types::Transaction) -> Self {
         let status = get_transaction_status(&tx.metadata);
         let ptype = get_payment_type(&tx.metadata);
-        let btc_tx_id = get_btc_tx_id(&tx.metadata);
+        let alpha_btc_tx_id = get_btc_alpha_tx_id(&tx.metadata);
+        let beta_btc_tx_id = get_btc_beta_tx_id(&tx.metadata);
         Self {
             id: tx.id().to_string(),
             amount: u64::from(tx.amount),
@@ -807,7 +816,10 @@ impl std::convert::From<cdk::wallet::types::Transaction> for Transaction {
             memo: tx.memo,
             ptype: ptype.into(),
             status: status.into(),
-            btc_tx_id: btc_tx_id.map(|id| id.to_string()),
+            melt_tx: MeltTx {
+                alpha_tx_id: alpha_btc_tx_id.map(|txid| txid.to_string()),
+                beta_tx_id: beta_btc_tx_id.map(|txid| txid.to_string()),
+            },
             quote_id: tx.quote_id,
         }
     }